refactor: split convert-row-and-table.ts into separate files
Split the convert-row-and-table.ts file into three separate files, each containing a single function:
- convert-array-of-rows-to-table-node.ts
- convert-table-node-to-array-of-rows.ts
- move-row-in-array-of-rows.ts

Also split the corresponding test file into three separate test files with matching names.

Updated import statements in move-column.ts and move-row.ts to reference the new individual files.
