+++ packages/extensions/src/table/table-utils/move-row.ts
@@ -5,11 +5,9 @@ import {
   TableMap,
 } from 'prosemirror-tables'
 
-import {
-  convertArrayOfRowsToTableNode,
-  convertTableNodeToArrayOfRows,
-  moveRowInArrayOfRows,
-} from './convert-row-and-table'
+import { convertArrayOfRowsToTableNode } from './convert-array-of-rows-to-table-node'
+import { convertTableNodeToArrayOfRows } from './convert-table-node-to-array-of-rows'
+import { moveRowInArrayOfRows } from './move-row-in-array-of-rows'
 import { getSelectionRangeInRow } from './get-selection-range-in-row'
 import { findTable } from './query'
